+++ src/main/java/com/zx/sms/codec/cmpp/msg/CmppActiveTestResponseMessage.java
@@ -13,7 +13,7 @@ import com.zx.sms.codec.cmpp.packet.CmppPacketType;
 public class CmppActiveTestResponseMessage extends DefaultMessage{
 	private static final long serialVersionUID = 4300214238350805590L;
 	private short reserved = 0;
-	public CmppActiveTestResponseMessage(long sequenceId) {
+	public CmppActiveTestResponseMessage(int sequenceId) {
 		super(CmppPacketType.CMPPACTIVETESTRESPONSE,sequenceId);
 	}
 	public CmppActiveTestResponseMessage(Header header) {
