+++ src/main/java/com/zx/sms/codec/cmpp/msg/CmppCancelResponseMessage.java
@@ -14,7 +14,7 @@ public class CmppCancelResponseMessage extends DefaultMessage {
 	private static final long serialVersionUID = -1111862395776885021L;
 	private long successId = 1;
 	
-	public CmppCancelResponseMessage(long sequenceId) {
+	public CmppCancelResponseMessage(int sequenceId) {
 		super(CmppPacketType.CMPPCANCELRESPONSE,sequenceId);
 	}
