+++ src/main/java/com/zx/sms/codec/cmpp/msg/Header.java
@@ -14,10 +14,10 @@ public interface Header extends Serializable {
     public long getPacketLength();
     public void setBodyLength(long length);
     public long getBodyLength();
-    public void setCommandId(long commandId);
-    public long getCommandId();
-    public void setSequenceId(long transitionId);
-    public long getSequenceId();
+    public void setCommandId(int commandId);
+    public int getCommandId();
+    public void setSequenceId(int transitionId);
+    public int getSequenceId();
     public long getNodeId();
     public void setNodeId(long nodeId);
 }
