+++ src/main/java/com/zx/sms/codec/sgip12/codec/SgipDeliverRequestMessageCodec.java
@@ -47,7 +47,7 @@ public class SgipDeliverRequestMessageCodec extends MessageToMessageCodec<Messag
 	@Override
 	protected void decode(ChannelHandlerContext ctx, Message message, List<Object> out) throws Exception {
 
-		long commandId = ((Long) message.getHeader().getCommandId()).longValue();
+		int commandId =  message.getHeader().getCommandId();
 		if (packetType.getCommandId() != commandId) {
 			// 不解析，交给下一个codec
 			out.add(message);
