+++ src/main/java/com/zx/sms/codec/sgip12/codec/SgipMessageCodecAggregator.java
@@ -22,7 +22,7 @@ public class SgipMessageCodecAggregator extends ChannelDuplexHandler {
 		private final static  SgipMessageCodecAggregator instance = new SgipMessageCodecAggregator();
 	}
 	
-	private ConcurrentHashMap<Long, MessageToMessageCodec> codecMap = new ConcurrentHashMap<Long, MessageToMessageCodec>();
+	private ConcurrentHashMap<Integer, MessageToMessageCodec> codecMap = new ConcurrentHashMap<Integer, MessageToMessageCodec>();
 
 	private SgipMessageCodecAggregator() {
 		for (PacketType packetType : SgipPacketType.values()) {
@@ -37,7 +37,7 @@ public class SgipMessageCodecAggregator extends ChannelDuplexHandler {
 	@Override
 	public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
 
-		long commandId = (Long) ((Message) msg).getHeader().getCommandId();
+		int commandId =  ((Message) msg).getHeader().getCommandId();
 		MessageToMessageCodec codec = codecMap.get(commandId);
 		codec.channelRead(ctx, msg);
 	}
@@ -45,7 +45,7 @@ public class SgipMessageCodecAggregator extends ChannelDuplexHandler {
 	@Override
 	public void write(ChannelHandlerContext ctx, Object msg, ChannelPromise promise) throws Exception {
 		try {
-			long commandId = (Long) ((Message) msg).getHeader().getCommandId();
+			int commandId =  ((Message) msg).getHeader().getCommandId();
 			MessageToMessageCodec codec = codecMap.get(commandId);
 			codec.write(ctx, msg, promise);
 		}catch(Exception ex) {
